Refactor comment tree logic
Place the comment tree logic within the comment parser module. Not a
real logic change, just code base cleanliness.
