+++ src/parser/comments.rs
@@ -1,4 +1,5 @@
 use std::error::Error;
+use std::collections::VecDeque;
 use log;
 use scraper::Html;
 use scraper::ElementRef;
@@ -8,6 +9,8 @@ use crate::model::Comment;
 use crate::error::HnError;
 use crate::model::Id;
 use crate::parser::HtmlParse;
+
+const COMMENT_INDENT_INCR: u32 = 40;
 
 lazy_static! {
     // Applied to root of HTML document
@@ -159,5 +162,39 @@ impl CommentsParser {
 
         Ok(indent)
     }
+}
+
+pub fn create_comment_tree(comments: Vec<Comment>) -> Vec<Comment> {
+
+    #[allow(clippy::comparison_chain)]
+    fn _create_comment_tree(q: &mut VecDeque<Comment>, parent: &mut Comment) {
+        let mut last: Option<&mut Comment> = None;
+        while let Some(c) = q.front() {
+            if c.indent == parent.indent + COMMENT_INDENT_INCR {
+                let c = q.pop_front().unwrap();
+                parent.children.push(c);
+                last = Some(parent.children.last_mut().unwrap());
+            }
+            else if c.indent > parent.indent + COMMENT_INDENT_INCR {
+                let next_parent = last.take()
+                    .expect("Jumped a nesting level in comment node hierarchy");
+                _create_comment_tree(q, next_parent);
+            }
+            else {
+                return;
+            }
+        }
+    }
+
+    let mut q = VecDeque::from(comments);
+    let mut forest = Vec::new();
+
+    while let Some(root) = q.pop_front() {
+        forest.push(root);
+        let ptr = forest.last_mut().unwrap();
+        _create_comment_tree(&mut q, ptr);
+    }
+
+    forest
 }
